// src/api-spec/protobuf/gen/js/tdex/v1/trade_pb.d.ts
export interface ListMarketsRequest {
}
/**
 * @generated from protobuf message tdex.v1.ListMarketsResponse
 */
export interface ListMarketsResponse {
    /**
     * @generated from protobuf field: repeated tdex.v1.MarketWithFee markets = 1;
     */
    markets: MarketWithFee[];
}
/**
 * @generated from protobuf message tdex.v1.GetMarketBalanceRequest
 */
export interface GetMarketBalanceRequest {
    /**
     * @generated from protobuf field: tdex.v1.Market market = 1;
     */
    market?: Market;
}
// ...
export interface PreviewTradeRequest {
    /**
     * @generated from protobuf field: tdex.v1.Market market = 1;
     */
    market?: Market;
    /**
     * @generated from protobuf field: tdex.v1.TradeType type = 2;
     */
    type: TradeType;
    /**
     * @generated from protobuf field: uint64 amount = 3;
     */
    amount: string;
    /**
     * @generated from protobuf field: string asset = 4;
     */
    asset: string;
}
/**
 * @generated from protobuf message tdex.v1.PreviewTradeResponse
 */
export interface PreviewTradeResponse {
    /**
     * @generated from protobuf field: repeated tdex.v1.Preview previews = 1;
     */
    previews: Preview[];
}
/**
 * @generated from protobuf message tdex.v1.ProposeTradeRequest
 */
export interface ProposeTradeRequest {
    /**
     * @generated from protobuf field: tdex.v1.Market market = 1;
     */
    market?: Market;
    /**
     * @generated from protobuf field: tdex.v1.TradeType type = 2;
     */
    type: TradeType;
    /**
     * @generated from protobuf field: tdex.v1.SwapRequest swap_request = 3;
     */
    swapRequest?: SwapRequest;
}
// ...
export interface CompleteTradeRequest {
    /**
     * @generated from protobuf field: tdex.v1.SwapComplete swap_complete = 1;
     */
    swapComplete?: SwapComplete;
    /**
     * @generated from protobuf field: tdex.v1.SwapFail swap_fail = 2;
     */
    swapFail?: SwapFail;
}
/**
 * @generated from protobuf message tdex.v1.CompleteTradeResponse
 */
export interface CompleteTradeResponse {
    /**
     * @generated from protobuf field: string txid = 1;
     */
    txid: string;
    /**
     * @generated from protobuf field: tdex.v1.SwapFail swap_fail = 2;
     */
    swapFail?: SwapFail;
}
declare class ListMarketsRequest$Type extends MessageType<ListMarketsRequest> {
    constructor();
    create(value?: PartialMessage<ListMarketsRequest>): ListMarketsRequest;
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: ListMarketsRequest): ListMarketsRequest;
    internalBinaryWrite(message: ListMarketsRequest, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter;
}
/**
 * @generated MessageType for protobuf message tdex.v1.ListMarketsRequest
 */
export declare const ListMarketsRequest: ListMarketsRequest$Type;
// ...
declare class GetMarketBalanceRequest$Type extends MessageType<GetMarketBalanceRequest> {
    constructor();
    create(value?: PartialMessage<GetMarketBalanceRequest>): GetMarketBalanceRequest;
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: GetMarketBalanceRequest): GetMarketBalanceRequest;
    internalBinaryWrite(message: GetMarketBalanceRequest, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter;
}
/**
 * @generated MessageType for protobuf message tdex.v1.GetMarketBalanceRequest
 */
export declare const GetMarketBalanceRequest: GetMarketBalanceRequest$Type;
// ...
declare class PreviewTradeRequest$Type extends MessageType<PreviewTradeRequest> {
    constructor();
    create(value?: PartialMessage<PreviewTradeRequest>): PreviewTradeRequest;
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: PreviewTradeRequest): PreviewTradeRequest;
    internalBinaryWrite(message: PreviewTradeRequest, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter;
}
/**
 * @generated MessageType for protobuf message tdex.v1.PreviewTradeRequest
 */
export declare const PreviewTradeRequest: PreviewTradeRequest$Type;
// ...
declare class ProposeTradeRequest$Type extends MessageType<ProposeTradeRequest> {
    constructor();
    create(value?: PartialMessage<ProposeTradeRequest>): ProposeTradeRequest;
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: ProposeTradeRequest): ProposeTradeRequest;
    internalBinaryWrite(message: ProposeTradeRequest, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter;
}
/**
 * @generated MessageType for protobuf message tdex.v1.ProposeTradeRequest
 */
export declare const ProposeTradeRequest: ProposeTradeRequest$Type;
// ...
declare class CompleteTradeRequest$Type extends MessageType<CompleteTradeRequest> {
    constructor();
    create(value?: PartialMessage<CompleteTradeRequest>): CompleteTradeRequest;
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: CompleteTradeRequest): CompleteTradeRequest;
    internalBinaryWrite(message: CompleteTradeRequest, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter;
}
/**
 * @generated MessageType for protobuf message tdex.v1.CompleteTradeRequest
 */
export declare const CompleteTradeRequest: CompleteTradeRequest$Type;